requirements.txt
alembic==1.4.1
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dominate==2.5.1
elasticsearch==7.6.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-Threads==0.0.1a0
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
urllib3==1.25.8
visitor==0.1.3
Werkzeug==1.0.0
WTForms==2.2.1
